refac: command Definition has options functionality
